requirements.txt
pandas
numpy
tqdm
argparse
python-dotenv
#torch
#sklearn
psutil
#SQLAlchemy
shutil
Pillow